Промежуточный коммит рефакторинга боевых умений. Каждому умению отдельный файлик
