+++ src/char.hpp
@@ -653,7 +653,7 @@ public:
 	void cleanup_script();
 
 	bool is_npc() const { return char_specials.saved.act.get(MOB_ISNPC); }
-
+    bool have_mind();
 private:
 	const auto& get_player_specials() const { return player_specials; }
 	auto& get_player_specials() { return player_specials; }
