+++ src/interpreter.cpp
@@ -15,6 +15,22 @@
 
 #include "interpreter.h"
 
+#include "skills/do_flee.h"
+#include "skills/do_bash.h"
+#include "skills/do_stun.h"
+#include "skills/do_resque.h"
+#include "skills/do_kick.h"
+#include "skills/do_strangle.h"
+#include "skills/do_chopoff.h"
+#include "skills/do_disarm.h"
+#include "skills/do_stupor.h"
+#include "skills/do_ironwind.h"
+#include "skills/do_throw.h"
+#include "skills/do_mighthit.h"
+#include "skills/do_block.h"
+#include "skills/do_parry.h"
+#include "skills/do_protect.h"
+
 #include "act.movement.hpp"
 #include "world.characters.hpp"
 #include "object.prototypes.hpp"
@@ -169,7 +185,6 @@ void do_at(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_affects(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_backstab(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_ban(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_bash(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_beep(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_cast(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_warcry(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -203,7 +218,6 @@ void do_remort(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_remember_char(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_exit(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_exits(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_flee(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horseon(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horseoff(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_horseput(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -235,14 +249,11 @@ void do_inspect(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_insult(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_inventory(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_invis(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_kick(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_kill(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_last(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_mode(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_mark(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_makefood(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_disarm(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_chopoff(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_deviate(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_levels(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_liblist(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -275,7 +286,6 @@ void do_rent(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_reply(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_report(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_refill(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_rescue(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_stopfight(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_setall(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_stophorse(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -301,7 +311,6 @@ void do_stat(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_steal(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_switch(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_syslog(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_throw(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_teleport(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_tell(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_time(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -326,29 +335,21 @@ void do_wiznet(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_wizutil(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_write(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_zreset(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_parry(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_multyparry(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_style(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_poisoned(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_repair(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_camouflage(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_stupor(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_mighthit(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_block(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_touch(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_transform_weapon(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_protect(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_dig(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_insertgem(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_ignore(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_proxy(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_turn_undead(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_iron_wind(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_exchange(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_godtest(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_print_armor(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_relocate(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_strangle(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_custom_label(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_quest(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_check(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -400,7 +401,6 @@ void do_check_occupation(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_delete_obj(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_arena_restore(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void Bonus::do_bonus_info(CHAR_DATA*, char*, int, int);
-void do_stun(CHAR_DATA*, char*, int, int);
 void do_showzonestats(CHAR_DATA*, char*, int, int);
 void do_overstuff(CHAR_DATA *ch, char*, int, int);
 void do_cities(CHAR_DATA *ch, char*, int, int);
