+++ src/poison.cpp
@@ -17,9 +17,8 @@
 #include "skills.h"
 #include "room.hpp"
 #include "fight.h"
+#include "act.offensive.h"
 
-extern void drop_from_horse(CHAR_DATA *victim);
-extern void set_wait(CHAR_DATA * ch, int waittime, int victim_in_room);
 extern int interpolate(int min_value, int pulse);
 
 namespace
