+++ src/skills/do_backstab.h
@@ -1,0 +1,12 @@
+#ifndef BYLINS_DO_BACKSTAB_H
+#define BYLINS_DO_BACKSTAB_H
+
+#include "char.hpp"
+
+void go_backstab(CHAR_DATA * ch, CHAR_DATA * vict);
+void do_backstab(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_BACKSTAB_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
+
