+++ src/skills/do_bash.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_BASH_H
+#define BYLINS_DO_BASH_H
+
+#include "char.hpp"
+
+void do_bash(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+void go_bash(CHAR_DATA * ch, CHAR_DATA * vict);
+
+#endif //BYLINS_DO_BASH_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
