+++ src/skills/do_block.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_DO_BLOCK_H
+#define BYLINS_DO_BLOCK_H
+
+#include "char.hpp"
+
+void do_block(CHAR_DATA *ch, char* /*argument*/, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_BLOCK_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
