+++ src/skills/do_disarm.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_DISARM_H
+#define BYLINS_DO_DISARM_H
+
+#include "char.hpp"
+
+void go_disarm(CHAR_DATA * ch, CHAR_DATA * vict);
+void do_disarm(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_DISARM_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
