+++ src/skills/do_flee.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_FLEE_H
+#define BYLINS_DO_FLEE_H
+
+#include "char.hpp"
+
+void go_flee(CHAR_DATA* ch);
+
+#endif //BYLINS_DO_FLEE_H
+
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
