+++ src/skills/do_ironwind.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_DO_IRONWIND_H
+#define BYLINS_DO_IRONWIND_H
+
+#include "char.hpp"
+
+void do_iron_wind(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_IRONWIND_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp
