+++ src/skills/do_kick.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_KICK_H
+#define BYLINS_DO_KICK_H
+
+#include "char.hpp"
+
+void go_kick(CHAR_DATA * ch, CHAR_DATA * vict);
+void do_kick(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_KICK_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
