+++ src/skills/do_mighthit.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_MIGHTHIT_H
+#define BYLINS_DO_MIGHTHIT_H
+
+#include "char.hpp"
+
+void go_mighthit(CHAR_DATA * ch, CHAR_DATA * victim);
+void do_mighthit(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_DO_MIGHTHIT_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
