+++ src/skills/do_parry.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_PARRY_H
+#define BYLINS_DO_PARRY_H
+
+#include "char.hpp"
+
+void do_multyparry(CHAR_DATA *ch, char* /*argument*/, int/* cmd*/, int/* subcmd*/);
+void do_parry(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
+
+#endif //BYLINS_DO_PARRY_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
