+++ src/skills/do_protect.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_PROTECT_H
+#define BYLINS_DO_PROTECT_H
+
+#include "char.hpp"
+
+void go_protect(CHAR_DATA * ch, CHAR_DATA * vict);
+void do_protect(CHAR_DATA *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+CHAR_DATA *try_protect(CHAR_DATA * victim, CHAR_DATA * ch);
+
+
+#endif //BYLINS_DO_PROTECT_H
