+++ src/skills/do_stun.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_STUN_H
+#define BYLINS_DO_STUN_H
+
+#include "char.hpp"
+
+void go_stun(CHAR_DATA * ch, CHAR_DATA * vict);
+void do_stun(CHAR_DATA* ch, char* argument, int, int);
+
+#endif //BYLINS_DO_STUN_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
