+++ src/skills/do_throw.h
@@ -1,0 +1,11 @@
+#ifndef BYLINS_DO_THROW_H
+#define BYLINS_DO_THROW_H
+
+#include "char.hpp"
+
+void go_throw(CHAR_DATA * ch, CHAR_DATA * victim);
+
+
+#endif //BYLINS_DO_THROW_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
